Macro hygiene for constant generics has been fixed
